Add Ctrl+Enter submission to Textarea and move contact form to dedicated folder
