+++ src/app/contact/page.tsx
@@ -1,4 +1,4 @@
-import { ContactForm } from '@/components/contact-form';
+import { ContactForm } from '@/components/contact/contact-form';
 
 export default function ContactPage() {
   return (
