+++ src/components/contact/contact-form.tsx
@@ -4,10 +4,10 @@ import { submitContactForm } from '@/lib/actions/contactForm';
 import { contactFormSchema, ContactFormState } from '@/lib/schemas/contactForm';
 import { zodResolver } from '@hookform/resolvers/zod';
 import { MailCheck } from 'lucide-react';
-import { startTransition, useActionState } from 'react';
+import React, { startTransition, useActionState, useRef } from 'react';
 import { useForm } from 'react-hook-form';
 import { z } from 'zod';
-import { Button } from './ui/button';
+import { Button } from '../ui/button';
 import {
   Form,
   FormControl,
@@ -15,9 +15,9 @@ import {
   FormItem,
   FormLabel,
   FormMessage,
-} from './ui/form';
-import { Input } from './ui/input';
-import { Textarea } from './ui/textarea';
+} from '../ui/form';
+import { Input } from '../ui/input';
+import { Textarea } from '../ui/textarea';
 
 const initialState: ContactFormState = {
   success: false,
@@ -29,6 +29,7 @@ export function ContactForm() {
     submitContactForm,
     initialState
   );
+  const formRef = useRef<HTMLFormElement>(null);
 
   const form = useForm<z.infer<typeof contactFormSchema>>({
     resolver: zodResolver(contactFormSchema),
@@ -44,6 +45,19 @@ export function ContactForm() {
     startTransition(() => {
       formAction({ reset: true });
     });
+  };
+
+  const handleKeyDown = (e: React.KeyboardEvent<HTMLTextAreaElement>) => {
+    if ((e.ctrlKey || e.metaKey) && e.key === 'Enter') {
+      e.preventDefault();
+      if (formRef.current) {
+        const submitEvent = new Event('submit', {
+          bubbles: true,
+          cancelable: true,
+        });
+        formRef.current.dispatchEvent(submitEvent);
+      }
+    }
   };
 
   if (state.success) {
@@ -70,7 +84,7 @@ export function ContactForm() {
   return (
     <Form {...form}>
       <h1 className="font-bold text-center text-xl pb-8">Get in touch!</h1>
-      <form action={formAction} className="space-y-8">
+      <form action={formAction} ref={formRef} className="space-y-8">
         <FormField
           control={form.control}
           name="email"
@@ -78,7 +92,11 @@ export function ContactForm() {
             <FormItem>
               <FormLabel>Email</FormLabel>
               <FormControl>
-                <Input placeholder="example@domain.com" {...field}></Input>
+                <Input
+                  placeholder="example@domain.com"
+                  disabled={isPending}
+                  {...field}
+                ></Input>
               </FormControl>
               {!state.success && <FormMessage>{state.error.email}</FormMessage>}
             </FormItem>
@@ -93,6 +111,7 @@ export function ContactForm() {
               <FormControl>
                 <Input
                   placeholder="Let us know what your query is about..."
+                  disabled={isPending}
                   {...field}
                 ></Input>
               </FormControl>
@@ -111,6 +130,8 @@ export function ContactForm() {
               <FormControl>
                 <Textarea
                   placeholder="...then hit submit and we'll get back to you as soon as we can!"
+                  disabled={isPending}
+                  onKeyDown={handleKeyDown}
                   className="min-h-24"
                   {...field}
                 ></Textarea>
@@ -123,7 +144,7 @@ export function ContactForm() {
         />
         <div className="flex gap-12 items-center">
           <Button type="submit" disabled={isPending}>
-            {isPending ? 'Submitting' : state.success ? 'Submitted!' : 'Submit'}
+            {isPending ? 'Submitting' : 'Submit'}
           </Button>
         </div>
       </form>
